+++ src/core/mod.rs
@@ -4,3 +4,4 @@ pub mod parser;
 pub mod project;
 pub mod resource_parser;
 pub mod scene;
+pub mod symbol_table;